requirements.txt
amqp==5.1.0
arabic-reshaper==2.1.3
asgiref==3.5.0
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.3
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==4.0.1
django-celery-results==2.3.0
django-cors-headers==3.11.0
django-filter==21.1
django-logpipe==1.1.0
django-timezone-field==4.2.3
djangorestframework==3.13.1
future==0.18.2
html5lib==1.1
importlib-metadata==4.10.1
kafka-python==2.0.2
kombu==5.2.4
lru-dict==1.1.7
Markdown==3.3.6
packaging==21.3
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.3
PyJWT==2.3.0
pyparsing==3.0.8
PyPDF2==1.26.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2021.3
redis==4.2.2
reportlab==3.6.6
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.0
xhtml2pdf==0.2.5
zipp==3.7.0
